% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_db_readblast.R
\name{get.fungal.palette.unite}
\alias{get.fungal.palette.unite}
\title{Creates a color palette based on fungal taxonomy}
\usage{
get.fungal.palette.unite(tax, parapsilosis = FALSE)
}
\arguments{
\item{tax}{either a data.frame, phyloseq, or tax_table}

\item{parapsilosis}{logical, if \code{FALSE} (default), C. parapsilosis will only have one color, if \code{TRUE}, will color according to C. parapsilosis ASV}
}
\description{
This function is adapted from \code{yingtools2::get.yt.palette} . It allows the use of the UNITE taxonomy (which differs from NCBI taxonomy).
}
\author{
Thierry Rolling
}
