% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_db_readblast.R
\name{read.blastn.unite}
\alias{read.blastn.unite}
\title{Reads in a UNITE taxonomy table}
\usage{
read.blastn.unite(tax.file = "asv_seqs.fasta.unite.txt", tax_table = TRUE)
}
\arguments{
\item{tax.file}{The output file from blasting the ASV table to the UNITE database. Default is \code{"asv_seqs.fasta.unite.txt"}}

\item{tax_table}{logical, if \code{TRUE} (default), will return a data frame of taxonomy with top hit, ready to be converted to a tax_table, if \code{FALSE}, returns all hits}
}
\value{
Taxonomy table data frame
}
\description{
This function is adapted from \code{yingtools2::read.blastn.file} . It allows reading in a taxonomic annotation file blasted to UNITE.
The top hit is chosen first by evalue (lowest evalue), if several hits with same evalue, the Species with the highest
number of hits is chosen.
}
\author{
Thierry Rolling
}
